Add tests for LUN parsing, simplify conf for the snapshot provider
